More render refactoring, proper mesh materials (almost)
Material refactor (still not happy with textures)
 BUT removed all hardcoded values + allows easy scaling.
Removed sort keys, added distance based sorting (front to back)
Some small refactors

+++ playground/engine/private/render/vertex_array.cpp
@@ -10,7 +10,7 @@ Matt Hoyle
 namespace Render
 {
 	// this must match the enum VertexDataType
-	constexpr size_t VertexDataTypeSizes[] = {
+	constexpr uint32_t VertexDataTypeSizes[] = {
 		sizeof(float)
 	};
